refactor: remove api schema
