+++ apps/web/src/common/modules/navigations/gnb/modules/gnb-noti/modules/GNBNoticeTab.vue
@@ -74,7 +74,7 @@ import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 
 import { SpaceRouter } from '@/router';
-import type { TimeStamp } from '@/schema/common/model';
+import type { TimeStamp } from '@/schema/_common/model';
 import { store } from '@/store';
 import { i18n } from '@/translations';
