+++ apps/web/src/services/administration/iam/role/type.ts
@@ -3,7 +3,7 @@ import type { TranslateResult } from 'vue-i18n';
 import type { DefinitionField } from '@spaceone/design-system/types/data-display/tables/definition-table/type';
 
 
-import type { Tags, TimeStamp } from '@/schema/common/model';
+import type { Tags, TimeStamp } from '@/schema/_common/model';
 
 import type { PagePermissionType, RawPagePermission } from '@/lib/access-control/config';
