+++ apps/web/src/services/administration/iam/user/type.ts
@@ -1,7 +1,7 @@
 import VueI18n from 'vue-i18n';
 import type { Location } from 'vue-router';
 
-import type { Tags, TimeStamp } from '@/schema/common/model';
+import type { Tags, TimeStamp } from '@/schema/_common/model';
 
 import type { RoleBindingType } from '@/services/administration/type';
