+++ apps/web/src/services/asset-inventory/cloud-service/cloud-service-detail/modules/cloud-service-history-detail/CloudServiceMultipleSelectedData.vue
@@ -17,8 +17,8 @@ import { QueryHelper } from '@cloudforet/core-lib/query';
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 
-import { QueryType } from '@/schema/common/api-verbs/export';
-import type { ExportParameter } from '@/schema/common/api-verbs/export';
+import { QueryType } from '@/schema/_common/api-verbs/export';
+import type { ExportParameter } from '@/schema/_common/api-verbs/export';
 import { store } from '@/store';
 import { i18n } from '@/translations';
