+++ apps/web/src/services/asset-inventory/service-account/type.ts
@@ -1,7 +1,7 @@
 import type { JsonSchema } from '@spaceone/design-system/types/inputs/forms/json-schema-form/type';
 
 
-import type { Tags, TimeStamp } from '@/schema/common/model';
+import type { Tags, TimeStamp } from '@/schema/_common/model';
 
 import type { Tag } from '@/common/components/forms/tags-input-group/type';
