+++ apps/web/src/services/project/types/type.ts
@@ -1,6 +1,6 @@
 import type { TreeItem, Tree } from '@spaceone/design-system/types/data-display/tree/type';
 
-import type { TimeStamp } from '@/schema/common/model';
+import type { TimeStamp } from '@/schema/_common/model';
 
 export type ItemType = 'PROJECT_GROUP'|'PROJECT';
